C API: refactor ListBuilder
