+++ src/libexpr/c/nix_api_expr_internal.h
@@ -3,6 +3,33 @@
 
 #include "eval.hh"
 #include "attr-set.hh"
+#include "nix_api_value.h"
+
+class CListBuilder
+{
+private:
+    std::vector<nix::Value *> values;
+
+public:
+    CListBuilder(size_t capacity)
+    {
+        values.reserve(capacity);
+    }
+
+    void push_back(nix::Value * value)
+    {
+        values.push_back(value);
+    }
+
+    Value * finish(nix::EvalState * state, nix::Value * list)
+    {
+        state->mkList(*list, values.size());
+        for (size_t n = 0; n < list->listSize(); ++n) {
+            list->listElems()[n] = values[n];
+        }
+        return list;
+    }
+};
 
 struct EvalState
 {
@@ -12,6 +39,11 @@ struct EvalState
 struct BindingsBuilder
 {
     nix::BindingsBuilder builder;
+};
+
+struct ListBuilder
+{
+    CListBuilder builder;
 };
 
 struct nix_string_return
